# backend/src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Digital Asset Warranty Platform API
  description: API for managing digital assets and warranty quotes
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local development server
paths:
  /assets:
    post:
      summary: Create a new digital asset
      description: Create a new digital asset with specified details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - category
                - value
                - purchaseDate
              properties:
                name:
                  type: string
                  description: Name of the digital asset
                category:
                  type: string
                  description: Category of the digital asset (e.g., electronics, watches, collectibles)
                value:
                  type: number
                  description: Value of the digital asset
                purchaseDate:
                  type: string
                  format: date
                  description: Purchase date of the digital asset
      responses:
        '201':
          description: Asset created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  category:
                    type: string
                  value:
                    type: number
                  purchaseDate:
                    type: string
                    format: date
                  createdAt:
                    type: string
                    format: date-time
                  updatedAt:
                    type: string
                    format: date-time
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    get:
      summary: Get all digital assets
      description: Retrieve a list of all registered digital assets
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    category:
                      type: string
                    value:
                      type: number
                    purchaseDate:
                      type: string
                      format: date
                    createdAt:
                      type: string
                      format: date-time
                    updatedAt:
                      type: string
                      format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /assets/{id}/quote:
    get:
      summary: Get warranty quote for an asset
      description: Get a simulated warranty quote for a specific digital asset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the digital asset
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  assetId:
                    type: integer
                  quoteAmount:
                    type: number
                  providerName:
                    type: string
                  validUntil:
                    type: string
                    format: date
        '404':
          description: Asset not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string